+++ src/spacecraftelements/Bullets/Bullet.java
@@ -15,23 +15,23 @@ public abstract class Bullet implements Updatable{
 	/**
 	 * 		子弹速度
 	 */
-	public Double vx,vy;
+	public int vx,vy;
 	/**
 	 * 		子弹加速度（可选）
 	 */
-	public Double ax, ay;
+	public int ax, ay;
 	/**
 	 * 		子弹初始生成位置
 	 */
-	public Double x, y;
+	public int x, y;
 	/**
 	 *		子弹朝向角度 	
 	 */
-	public Double angle;
+	public int angle;
 	/**
 	 * 		子弹伤害值
 	 */
-	public Double damage;
+	public int damage;
 	/**
 	 * 		子弹所属（跟踪类武器需要）
 	 * 		用于子弹判断敌我
@@ -40,6 +40,6 @@ public abstract class Bullet implements Updatable{
 	/**
 	 * 		子弹体积
 	 */
-	public Double volume;
+	public int volume;
 	public abstract boolean update();
 }
